On branch file-by-file-type-in-type
Here we change our basic approach: instead of trying to correct Coq's
optimization of universe level variables that have lower bounds, we try to
write code that is agnostic about comparisons and introduces no lower bounds.

For example, instead of

    Definition hsubtype@{i} (X : Type@{i}) : Type@{i} := X -> hProp.

we have

   Definition hsubtype@{i k} (X : Type@{i}) : Type@{k} := X -> hProp.

The old definition led to an inequality uu1 <= i, which we want to avoid.